// frontend/src/app/shared/models/cidade.model.ts
export class CidadeModel {
	distritoId: number;
	distritoNome: string;
	municipioId: number;
	municipioNome: string;
	microrregiaoId: number;
	microrregiaoNome: string;
	mesorregiaoId: number;
	mesorregiaoNome: string;
	regiaoImediataId: number;
	regiaoImediataNome: string;
	regiaoIntermediariaId: number;
	regiaoIntermediariaNome: string;
	ufId: number;
	ufSigla: string;
	ufNome: string;
	regiaoId: number;
	regiaoSigla: string;
	regiaoNome: string;
}